Emperors Word Updates
All of the traversal elements are there, you can now navigate from the start to the end. Next up is checkpoints placements, enemy placements and boss placements.  There are some enemies in place right now, but these are subject to change. Average traversal time is coming out at about 3.5 mins (That is with map knowledge and taking all shortcuts).